Implement correct scoring circuit, add tests
Implements a scoring circuit with the correct logic for handling
multiple occurrances of the same letter in the guess. That is, if a
letter appears `n` times in the word, then it should be scored at most
`n` times, even if it appears more than `n` times in the guess,
prioritizing green matches first and then matching yellows from left to
right.

Also add unit tests for this circuit.