Use updated signer to fix Safari private browsing
The new version of Alchemy Signer lets us receive the org id in the url
and manually pass it in to `signer.authenticate`, rather than relying on
local storage. This fixes login on Safari private browsing (both iOS and
desktop), which does not share local storage between tabs.

Also adds a temporary fix for improper escaping on the URL sent in the
email.